Optimize TRY_CAST(varchar as integer) for invalid strings. (#9872)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9872

Throwing exceptions is expensive. The new no-throw implementation is 20x
faster.

Replaced `T Converter<TypeKind>::cast(v)` API with non-throwing 'tryCast' that
returns `folly::Expected<T, Status>`

```
folly::Expected<T, Status> tryCast(v);
```

The implementation of Converter::cast used folly::to in many places. Replaced
most of them with non-throwing folly::tryTo. The only cases where we continue
to use folly::to is when converting to string. folly::to doesn't throw when
converting to std::string and folly::tryTo doesn't support converting to
string. This implies that `Converter<TypeKind::VARCHAR>::tryCast(v)` never
returns an  error and therefore it is safe to call '.value()' on the result
unconditionally.

The new Convert::tryCast API allows to avoid throwing exceptions in TRY and
TRY_CAST, but it uses Status object which is not cheap to create or copy. It
still creates error messages that are not used. We can optimize further by using
some sort of error code enum instead of Status. This could improve the
performance 2-3x further.

Note that in a few places we want to get the result of the 'cast' or throw. Such
code could benefit from Status::throw() helper function.

```
    const auto output =
        util::Converter<ToKind>::tryCast(simpleInput->valueAt(row))
            .thenOrThrow(folly::identity, [&](const Status& status) {
              VELOX_USER_FAIL(status.message());
            });
```

Switching to folly::tryTo has some impact on the quality of the error messages.
For example, before this change, `cast('1a' as bigint)` failed with error
message

```
Cannot cast VARCHAR '1a' to BIGINT. Non-whitespace character found after end of conversion: "a"
```

After this change the error message is

```
Cannot cast VARCHAR '1a' to BIGINT. Non-whitespace character found after end of conversion: ""
```

Before this change the error message included the tail of the string that failed
to parse ("a") and after this change the error message includes empty string
(""). folly::tryTo returns an error code in the form of ConversionCode enum. We
use `folly::makeConversionError(conversionCode, "")` to convert error code into
error message. This API takes ConversionCode enum and a string that contains
additional context to include in the error message. That context depends on
particular error code. We pass empty string for context because figuring out
what context to pass for which error code would couple our code with folly's a
bit too much.

Another example of different error message is for cast(integer '1111111' as tinyint).

Before:

```
Cannot cast INTEGER '1111111' to TINYINT. Overflow during arithmetic conversion: (signed char) 1111111
```

After:

```
Cannot cast INTEGER '1111111' to TINYINT. Overflow during arithmetic conversion:
```

Arguably, in this case we haven't lost any details. The lost tail "(signed char)
1111111" contained redundant information that's already present at the start of
the message: "Cannot cast INTEGER '1111111' to TINYINT". One may argue that the
original message was confusing because it referred to "signed char" type
instead of TINYINT. The new message seems complete and correct except for
ending abruptly with a colon and not a period.

Reviewed By: xiaoxmeng, Yuhta

Differential Revision: D57572110

fbshipit-source-id: c796ffd71b915d09cfb1edf6ec12df62ed666d01