Add APIs to get simple and vector function signatures in function registry (#9884)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9884

Currently we only have a way to get the signatures of all the functions that are registered into Velox - simple and vector. This utility API is generally useful to provide info about what UDFs are available to users.

In this change, I added 2 separate APIs to get simple and vector function signatures independently (replicating the same logic as the main one). This is helpful when the separation is desired.

Reviewed By: bikramSingh91

Differential Revision: D57647355

fbshipit-source-id: fad199f8afda6dd5b91275327f6c4981775cffdb