Add checksum for ssd cache (#9822)
Summary:
This PR introduces an optional checksum feature for the SSD cache. When
enabled, a CRC-based checksum is calculated for each cache entry and
stored in the checkpoint file. Additionally, if read verification is
activated, the checksum is recalculated and verified against the stored
value when cache data is recovered from the checkpoint or loaded from
the SSD.
A new counter, `ssd_cache_read_corruptions`, is added to track the
number of corruptions detected due to checksum mismatches.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9822

Reviewed By: xiaoxmeng

Differential Revision: D57394558

Pulled By: zacw7

fbshipit-source-id: 5a1ee22c11e63fe4fe4a3f7c67775ffcd942068e