Make IcebergQueryRunner and IcebergExternalQueryRunner DATA_DIR in sync
If the addStorageFormatToPath parameter is set to true, and $DATA_DIR
was given as a program parameter, then the actual data directory would
be "$DATA_DIR/$FILE_FORMAT" for IcebergExternalQueryRunner while just
"$DATA_DIR" for IcebergQueryRunner. This causes the data created by
one can not be seen and updated by the other. This commit set the
default value for addStorageFormatToPath to false, so that the data
can be shared between the two QueryRunner's.