Fix NestedLoopJoin with empty build or probe with no join condition (#9892)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9892

Join with no condition is a cross product. The existing code avoid
adding mismatch to the result after cross product because cross
product should have matched all input rows (https://github.com/facebookincubator/velox/pull/6010). But there
is an exception. When the build or probe side is empty, this cross
product is empty too. Hence for left, right, and full join, mismatch
should still be produced. This diff fixes this bug by still adding the
mismatch to the result if either build or probe side is empty.

Reviewed By: Yuhta

Differential Revision: D57681090

fbshipit-source-id: ac5960faf33166c4660bba25d516a2ffc1b6276c