Fix flaky HashJoinTest.failedToReclaimFromHashJoinBuildersInNonReclai (#9962)
Summary:
The test is flaky because of the following race condition. When reclaim happens in join node, there is a chance probe side pipeline/driver has not been initialized, making reclaimer not set for the operator. This results in probe side not reporting non reclaimable attempts because the reclaim codepath is not executed. Added additional coordination of code execution to avoid such inconsistent behavior of the test.

Stress tests have been run to verify the consistency of the fixed test.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9962

Reviewed By: xiaoxmeng

Differential Revision: D57880512

Pulled By: tanjialiang

fbshipit-source-id: bcc829301911f151d51aa68e8b3c75a93ebf5293