Add nan definition field to DoubleType and RealType
Add a boolean field to DoubleType and RealType to determine whether to
use the new nan definition.  If useNewNanDefinition is set to true in
the configuration property, then only doubles/reals with that property
set to true will be created, and if it is false, then only doubles/reals
with that property set to false will be created.  This will be used in
later commits to make decisions about how to handle nans.

Because DoubleType and RealType are now parametrized types, it is no
longer correct to use type == DOUBLE for type checking, as it is no
longer a singleton instance.  All code that was using type == DOUBLE or
type == REAL has been updated to use .equals() comparison