Replace ErrorVector with EvalErrors class (#9988)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9988

ErrorVector was a flat vector storing `std::shared_ptr<std::exception_ptr>`.

In many cases (e.g. TRY) it is sufficient to simply indicate that a row has an
error. It is not necessary to store the actual error that occurred. ErrorVector
used 'nulls' buffer to indicate whether a row has an error and 'values' buffer
to store the exception or null (if error details are not provided). Allocating
unnecessary 'values' buffer wastes ~20 bytes per row.

To use memory more efficiently, we replace ErrorVector with a custom
class EvalErrors. Initially, EvalErrors simply wraps ErrorVector. A follow-up
change will modify EvalErrors to store 'nulls' buffer and an optional 'values'
buffer directly. EvalErrors will not allocate 'values' buffer unless it is needed
to store error details.

A two-step transition is necessary to keep each change relatively small,
easier to review and safer to land.

Reviewed By: bikramSingh91

Differential Revision: D57992922

fbshipit-source-id: a60b8967f15cd4b9729019c7aa3929b8dc32bfa7