Vectorizing merge join (#9763)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9763

Changing the merge join algorithm to avoid the copy to a new output
buffer and rather wrapping the input left and right using dictionaries. Left
buffers are always wrapped, and right are wrapped unless the output spans more
than one right buffer (in which case it gets flattened).
.
The new version shows a ~10x speed up over the existing code, and about 4x speed up in a real end to end pipeline.

Meta:
The algorithm is described in detail in the document:
https://docs.google.com/document/d/1Yymt0DhzFUk4wwfaxd8fBuU9_bq_iNyBjKdlQxIURw0/edit#heading=h.6qpz4v5w6279

Profile before the change:

https://pxl.cl/4QGWv

Profile after the change:

https://pxl.cl/4QGWH

Reviewed By: mbasmanova

Differential Revision: D57146000

fbshipit-source-id: b406ab407204590783e3113a6d8d3c9c6d25bd40