Do not allocate dictionary filter cache if the filter is IsNotNull to avoid OOM (#10001)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10001

For map keys we added a `IsNotNull` filter to align the behavior with
Presto Java.  When the key column is dictionary encoded, we allocate the filter
cache even the dictionary values always pass the filter because they cannot be
null.  This caused some memory regression and we fix it by ignoring the
`IsNotNull` filter on dictionary values.

bypass-github-export-checks

Reviewed By: oerling

Differential Revision: D58023156

fbshipit-source-id: fca29bfbb1fa1de69dbc9c081b797bf7bf9e47c6