Fix crash in RowContainer::storeWithNulls<TypeKind::HUGEINT>
Summary:
Before:
When executing the `storeWithNulls` function on dictionary input, The `DecodedVector::valueAt` function will execute `index(idx)` to retrieve the corresponding index from `indices_`. However, if no null check is performed, the index is not guaranteed to be good, and performing offset calculation will result in a core dump.

After:
Check null values first, zero out the value if it is null.

Reviewed By: kevinwilfong

Differential Revision: D58087402

fbshipit-source-id: 862569866d3fdd60bce34723b9228b2333fefe12