Fix flaky reclaimFromHashJoinBuildInWaitForTableBuild (#10032)
Summary:
The test is flaky because there are two child pools from the query root pool. One is the
query task pool and the other is fake memory pool. And the test is designed to allocate all
the memory from the memory arbitrator and expect the memory arbitration reclaim all
the memory from the hash build operator after one of the hash build operator finishes
processing the input. And it expects the fake memory allocation succeed but the query
task fails with query OOM (the fake memory pool is not reclaimable). However the test is
recently found crashed in (github [issue](https://github.com/facebookincubator/velox/issues/10027)). The reason is because after reclaim from the hash join
query task, the task is resumed and there is a short time window that the hash build
operators start allocating more memory and before fake memory pool proceeds with the
rest of the memory arbitration process and the fake memory allocation fails with throw.

This PR deflaky this test by triggering the memory arbitration within the last the hash build
operator context to avoid the above flakiness. The meta internal test passed 1000 iterations:
https://www.internalfb.com/intern/testinfra/testrun/7318349614536715

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10032

Reviewed By: tanjialiang

Differential Revision: D58115774

Pulled By: xiaoxmeng

fbshipit-source-id: 32eb56b66ca923d81de2dff682f32eed4947550d