Use distinct semantics for set_agg and set_union
Use distinct semantics rather than equal to semantics for set_agg and
set_union.  This fixes an issue where arrays of rows or arrays that had
null fields/elements would fail with a message that comparison of rows
with null elements was not allowed.

This change required generating the set_agg and set_union function
directly rather than using annotation based generation because we needed
to add an argument to SetAggregationStateSerializer