Honor Presto's compression level configs and use the same defaults in HiveDataSink (#10117)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10117

Presto allows configuring the compression level used with ZSTD and ZLIB compressors when writing
files to Hive.  This change updates HiveDataSink to apply the configured value to the compression
level it  uses.

It also updates the default value(s) used in HiveDataSink to match Presto.  For ZLIB this is 4 and for
ZSTD this is 3.  This was a little awkward given that they're controlled by a shared single config in
Presto.

Reviewed By: xiaoxmeng

Differential Revision: D58313487

fbshipit-source-id: ca82c8bc0d43322d5f4e241b90dcc6bdc9142c10