Improve Fuzzer Job logging (#10119)
Summary:
Logging to stderr massively slows down the actual fuzzer jobs when a verbose log level is used but we still want the information. This PR adds logging to file (ERROR and above will still show on stderr) which will be uploaded as part of the repro artifacts.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10119

Reviewed By: pedroerp

Differential Revision: D58417649

Pulled By: kgpai

fbshipit-source-id: 9009d13da8e00511adbd2dee01003f085059354d