Fix writer OOM by reducing initial flush buffer reservation (#10147)
Summary:
Initial memory reservation is based on the tracked ratio maintained by RatioTracker. The one for flush is used for computing memory to reserve before flushing. For initial ratio it was default to 1.0f which ends up writer to reserve >2x (reservation headroom * ratio for flush) of raw memory from writes. This default ratio is used in the first iteration and is too large. Reduced it to 0.1 for better memory utilization. This helps to reduce in many cases writer memory usage by more than 60%.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10147

Reviewed By: xiaoxmeng

Differential Revision: D58440137

Pulled By: tanjialiang

fbshipit-source-id: c715f99e9a081cc1947cde4f2d1b9c9fbe020b50