Verify checksum for full SSD entry only (#10168)
Summary:
When loading data from 'ssdPins' into 'pins', mismatches occur if the requested 'pins' range is shorter than the SSD cache entries. This is because the checksum is calculated on partial data but compared with a checksum of the full entry. To resolve this, checksum verification should only be applied when reading the entire SSD entry.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10168

Reviewed By: xiaoxmeng

Differential Revision: D58503913

Pulled By: zacw7

fbshipit-source-id: ccbad747710e89ea9b3c9a91b951a55790d556e5