Fix timezone conversion bug in date_trunc (#10183)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10183

date_trunc was incorrectly handling truncations during the daylight
savings transition. Since conversions during these transitions may be
ambiguous, Presto semantic is to always take the earliest UTC timestamp.

For truncations that should happen in the "latest" side of an ambiguous
conversion, converting to UTC and back would result in the truncation being
done is the earliest version of that timestamp. The bug only applied to "minute"
and "hour" truncations, naturally.

Reviewed By: kevinwilfong, mbasmanova

Differential Revision: D58467148

fbshipit-source-id: dedaff4e599516a00c13dd9f372e6a6b547c2975