Fix wrong results bug with count over mixed aggregation
Fix a wrong results bug with count over an aggregation that had a mix of
global and non-global grouping sets.

We were not checking for single global aggregations correctly in
QueryCardinalityUtil, so we would return that the plan was scalar if
there were *any* empty grouping sets rather than if the empty grouping
set was the only grouping set.  we have now fixed this to return that
if all of the grouping sets are global, then the cardinality will be the
number of grouping sets, and otherwise it is at least the number of
global grouping sets.

This fixes queries like the following:
SELECT COUNT(*) FROM (SELECT count(*) FROM tpch.sf1.nation GROUP BY GROUPING SETS (nationkey, ()));

previously we would incorrectly return 1. And now we return 26.

This change may also fix bugs with correlated subqueries, as those also use
the isScalar() utility function.