Fix map_top_n on empty values. (#10207)
Summary:
Thanks to DanielHunte for pointing this out. Previous PR to fix map_top_n had a typo where it was missing a not , and thus when values were null it would not use the keys to compare. The test case would pass since it unfortunately had the higher value keys in ascending order and would pick the last  value. This PR fixes the bug and test case.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10207

Reviewed By: kagamiori

Differential Revision: D58623134

Pulled By: kgpai

fbshipit-source-id: 6fd2b2ff5bf7c7bcbcc19ca079e511c725f5ea5c