Fix autodeps for xsimd (#10198)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10198

Fix the autodeps for xsimd.  Currently auto deps tries to pull in pythran if a file depends on xsimd.
This is wrong and leads to broken builds if we just blindly accept it.

Fix autdeps to pull in xsimd from third-party if code depends on a header from xsimd.

Currently Velox is the only project that uses xsimd directly (some other projects use it as well, but
pull in Velox's header files to do so).

This also allows us to remove a buck directive from our open source Velox code.

Reviewed By: Yuhta

Differential Revision: D58592995

fbshipit-source-id: d0f9c9d114dfd7c3751413f8c5e501cde3432b20