Update clang-format lint rules (#10181)
Summary:
Now that we're on Ubuntu 24.04 we can use a more recent version of clang-format, and we can add rules that we run
within Meta but are missing in OS.  This should help to reduce back and forth between contributors and Meta employees
when importing their changes.

I also updated the script that does the lint checks to also apply the checks to renamed files, as those may have additional
changes besides just being renamed.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10181

Reviewed By: kgpai

Differential Revision: D58592253

Pulled By: kevinwilfong

fbshipit-source-id: 4d159caff85ceac6ebc6d2831ebdededed97a4be