Fix flaky test TableScanTest.statsBasedSkippingNulls (#10233)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10233

The test is flaky in optimized build on better hardware because the
minimum unit of our timer is 1 microsecond, so if one read is faster than it,
the scan time could be recorded as 0.  Remove the flaky check and beef up the
test with some stats it should check against.

Reviewed By: xiaoxmeng

Differential Revision: D58691228

fbshipit-source-id: 79ab929d9e7b7ebd7b9d8228061bfd224c1ddbc4