Add list of components and maintainers (#10180)
Summary:
This PR adds a webpage with a list of library components, their respective maintainers, and initial nomination for the different governance groups.

This is a follow up from the governance discussions started at VeloxCon'24. More context can be found in the document that describes the technical governance mechanics:

> https://velox-lib.io/docs/community/technical-governance

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10180

Reviewed By: Yuhta

Differential Revision: D58697120

Pulled By: pedroerp

fbshipit-source-id: b65295a58d9adc8a0180941248754d6c038dc4f5