Run preliminary checks in Docker container built for them (#10234)
Summary:
Update the preliminary checks GitHub workflow to use the check-avx Docker container we're already building.

The version of clang-format in whatever we're using now isn't up to date and is failing with some new rules I added.
Using the Docker image, which has a more recent version of clang-format, I verified it fixes the recent failures we're
seeing in running the "Code Format" job.

It also means we don't need to pip install dependencies in the jobs themselves anymore which is a small bonus.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10234

Reviewed By: kgpai

Differential Revision: D58702942

Pulled By: kevinwilfong

fbshipit-source-id: 3c74c978d3e332292eae1787a22cb3734d6eebe9