Fix memory cache hit underreporting in ioStats (#10272)
Summary:
Memory cache hits are only counted when the `isFirstUse` flag is set to false. However, when there's a cache miss and a new cache entry is created with `isFirstUse` set to true, the entry remains in exclusive mode and the `isFirstUse` flag stays true. Consequently, subsequent reads that hit this entry aren't counted as cache hits in ioStats, leading to underreporting. We need to ensure the flag is cleared properly after the entry is first loaded.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10272

Reviewed By: xiaoxmeng

Differential Revision: D58848317

Pulled By: zacw7

fbshipit-source-id: 1c17ed7c5bddf684c2ab0dd2a3f9bfab2e4ba227