Capture MemoryArbitrationContext and ThreadDebugInfo in AsyncSource and restore them when invoking make (#10186)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10186

In AsyncSource, at the time make is invoked, we've lost the MemoryArbitrationContext and
ThreadDebugInfo that were in scope at the time it was created.

To help provide debugging context, this diff modifies AsyncSource to capture and restore that info
so that it can be used if "make" should fail.

This was already being done for MemoryArbitrationContext inside of
createAsyncMemoryReclaimTask, this diff does the same thing in a more general way.

This was inspired by a crash that occurred in Prestissimo while spilling inside an AsyncSource, the
relevant debug info was lost.

Reviewed By: xiaoxmeng

Differential Revision: D58558331

fbshipit-source-id: 3599e770278e332f82e6bf2ba0172f29d8239b80