Add analyzer warning for common NaN pitfalls
Add a warning for comparisons on doubles and reals, where the semantics
signifcantly changed for NaN comparisons, and for division of doubles
and reals which commonly produces nans.  This warning is noisy, as it
will warn for all such operations on doubles and reals, and not just
those that involve nans.