Fix hash for double and real types when new nan is disabled
The hash function wasn't gated by the use-new-nan-definition field,
which meant it returned different results for -0 (which in the new
definition is the same as +0, and in the old definition was not).

This also changes the method call for the real hash function with
use-new-nan-definition enabled, but the two functions have the same
implementation, so the change is purely for clarity/consistency.