Add session timezone to Parquet PageReader (#9781)
Summary:
The Parquet PageReader needs to obtain the session timezone in order to support
INT64 Timestamps with isAdjustedToUTC=0. When processing it, the Timestamp needs
to be converted to UTC time based on the local timezone.

This PR provides access to session timezone to Parquet PageReader, but the reader
doesn't yet use this information.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9781

Reviewed By: Yuhta

Differential Revision: D58678144

Pulled By: kagamiori

fbshipit-source-id: d777b803d8ee7cbb4a5d581102edcce0e891d878