Add cmake helper to find libdouble-conversion (#10231)
Summary:
The package of libdouble-conversion-dev in Ubuntu 20.04 does not provide a cmake config file. Therefore, cmakes fails to initialize stating that it cannot find any of the names:
  double-conversionConfig.cmake,
  double-conversion-config.cmake.

This change provides a cmake helper to let Velox find the library. It was originally copied from:
  https://github.com/qt/qtbase/blob/v6.3.1/cmake/Finddouble-conversion.cmake
that was later refactored into:
  https://github.com/qt/qtbase/blob/dev/cmake/FindWrapSystemDoubleConversion.cmake

Fixes https://github.com/facebookincubator/velox/issues/10230

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10231

Reviewed By: amitkdutta

Differential Revision: D58825620

Pulled By: bikramSingh91

fbshipit-source-id: 910f7d85b6fa3f62ea2427a2e5ab7713f043a781