Add IO throttler support (#10285)
Summary:
This PR adds IO throttler support which can back off based on the storage throttling signal with a given file
path. Currently we support two kinds of throttling signals: one is local which only apply to a particular file folder
or directory, and the other is a storage cluster. Correspondingly, we expect two other inputs one is the file folder
path and the other is the cluster name which are actual storage system specific. Note here we assume that
a query system might access data from different storage systems and the cluster is the storage system abstraction.
The IO throttler takes a set of config parameters to control the actual backoff algorithms. Uses a cache to track
the recent storage throttling signals. We integrate this with Meta internal storage system to slow down the
IO when storage system is overloaded. Next we will add to support operation based throttling to prevent
server OOM from the non-velox memory usage by storage system.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10285

Reviewed By: bikramSingh91, tanjialiang, oerling

Differential Revision: D58900857

Pulled By: xiaoxmeng

fbshipit-source-id: da97ad2c9d762666dad3f203fb036ff9b8a1918d