Fix race condition in MemoryArbitrationFuzzer (#10314)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10314

MemoryArbitrationFuzzer reuses the same rng in multiple threads, this leads to TSAN reporting a
data race.  We can use a distinct rng in each thread with a seed generated by the original rng to
avoid this.

(The tests aren't entirely reproducible since they depend on the timing of the threads, so using
seeds from the original rng is not really necessary)

Reviewed By: xiaoxmeng

Differential Revision: D59031526

fbshipit-source-id: b50412aa3f420a0619a3782fa80eac726d47f0a5