Return predicate columns in table statistics
This happens with queries where the predicate
column isn't used in the output. It is pruned from the
assignments of the TableScanNode, but we need to
include the column in the TableStatistics output
so that the stats calculator can make an accurate
guess about the rows returned from the scan
with data from all predicate columns.