Speed up ApproxPercentileResultVerifier >100x (#10329)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10329

ApproxPercentileResultVerifier computes exact percentile ranges for each row and verifies the result
of approx_percentile is within some delta of that range.

The way it computes this when the function is called with a weight is to duplicate each row <weight>
times using a combination of the sequence UDF and unnest.  It then assigns row numbers to each
row after partitioning by the group by keys and sorting by the value column and computes the
minimum and maximum row numbers and divides it by the total number of rows to get the percentile
range.

The min row number is equivalent to summing up the weight column of all rows that precede it in the
partition, and the max is equivalent to adding the weight of the current row to the min.  This runs
much faster when weights are large (which is often the case in the fuzzer).

This addresses timeouts we've been seeing where a single query using approx_percentile can run
for over 15 minutes in debug builds or builds with TSAN enabled.

I also discovered and fixed a bug where we never verifying the first percentile.

Reviewed By: amitkdutta

Differential Revision: D59075897

fbshipit-source-id: 5e8834cb0b1e3064c35ac612d96fbefdfe9591d9