Don't run plans with LocalPartition in Aggregation and Window Fuzzers when TSAN is enabled (#10340)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10340

There is a known issue where LocalPartition will send DictionaryVectors with the same underlying base Vector to multiple threads.  This triggers TSAN to report data races, particularly if that base Vector is from the TableScan and reused.

To avoid these false negatives when the fuzzers are run with TSAN, I've modified the fuzzers to not
include plans that use LocalPartition when TSAN is enabled.  This was we still get TSAN coverage,
but without the noise.

Note that my initial attempt to ignore TSAN errors in WindowFuzzer does not work
https://github.com/facebookincubator/velox/pull/10315  I suspect that guard only disables TSAN
within the function it is initialized in, not the downstream scope.  I must have just gotten lucky when
testing it. This change also removes it.

Reviewed By: kagamiori

Differential Revision: D59122687

fbshipit-source-id: 0c0cf58819cb9c5154b33ec35ee61465f41cef07