Use property to deactivate UI profile
In lieu of the activeByDefault, use a property instead to
disable building the UI

activeByDefault doesn't quite do what the user might expect.
Other profiles which are activeByDefault can override this
in the root POM but not submodule POMs, so the the build
of the presto-main module might not be able to find the
presto-ui dependency.

This changes instead uses the <property> profile activation
which will only deactivate this profile in the case a specific
property (skipUI) is present. This can be passed to the maven
arguments as simply `-DskipUI`, similar to the surefire plugin's
`-DskipTests`

See the following for more discussion:
https://stackoverflow.com/questions/5309379/how-to-keep-maven-profiles-which-are-activebydefault-active-even-if-another-prof