Fix incorrect result of approx_percentile in window operations (#10368)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10368

approx_percentile(x, percentiles) returns incorrect result in a window
operation where a frame that should return non-NULL follows a frame
that returns NULL. Before the fix, the window operation returns NULL
for both frames. This is because
AggregationWindowFunction::computeAggregate() reuses the same
result vector across frames. It calls BaseVector::prepareForReuse()
before computing the aggregation for a frame. prepareForReuse()
doesn't clear the nulls buffer of the vector as it expects the code that
reuse the vector to set nulls properly. ApproxPercentileAggregate,
however, doesn't set the result rows to be non-null in the code path
for array-typed percentiles.

This diff fixes this bug by setting the nulls buffer properly.

Reviewed By: Yuhta

Differential Revision: D59257658

fbshipit-source-id: 1e114dad3a61401eb90292a4b40a662708dfe6a5