Check determinism using function name only (#10241)
Summary:
In ExpressionFuzzer, the determinism could be decided without argument
types. This PR changes to check determinism using function name only by
fetching all registry entries for the given function name and checking if all
of them are deterministic. The function is deterministic if "deterministic" is
true for every entry. With this change, argument types are not required when
the signature variables are not empty.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10241

Reviewed By: pedroerp

Differential Revision: D59122132

Pulled By: bikramSingh91

fbshipit-source-id: ab3c2907f102d3ebb81269fe58499bef0bb878ab