Use retainedSize instead of estimateFlatSize in LocalExchange (#10390)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10390

We observed some OOM happening because `LocalExchange` is buffering the
input vectors more than the memory limit.  These vectors contain dictionary with
very few indices but a large alphabet, i.e. result of filtering.  In
`LocalExchange` we use `estimateFlatSize` to calculate the memory usage of
buffered data, which only counts the rows selected, so the size is heavily
underestimated.  We fix this by using `retainedSize` instead of
`estimateFlatSize` in `LocalExchange` to count the memory usage more properly.

Reviewed By: xiaoxmeng, mbasmanova

Differential Revision: D59336200

fbshipit-source-id: f15c5f9d1d6f2cef22090088f5c7da5872951f54