Fix race condition in CacheFuzzer (#10415)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10415

The threads CacheFuzzer spins up use the loop variable "i" captured by reference as part of
generating the seed.  This loop variable changes on every iteration so we get concurrent reads and
writes to this variable which TSAN captures (and also probably violates the intention of the author).

Capturing a copy of the variable fixes it.

Reviewed By: pedroerp

Differential Revision: D59476752

fbshipit-source-id: c1a2aaa61e9dce8fb5d4e12ed762ed56a45f16e3