Relax ApproxDistinctResultVerifier for small number of groups with large error bound (#9748)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9748

When the number of groups is small, approx_distinct doesn't
guarantee the standard error. And when the input error bound is
large, the standard error is more likely to exceed the error bound. This
diff relaxes ApproxDistinctResultVerifier to not check the standard
error when the number of groups is below 50 and the error bound is
larger than the default 2.3%.

Reviewed By: Yuhta

Differential Revision: D57115412

fbshipit-source-id: ef4305aef46273ed5df3c068f586bd34c196de85