Respect Presto configs for enabling/disabling dictionary encoding in DWRF writer (#10428)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10428

Today we don't support the configs Presto uses to enable/disable dictionary encoding in DWRF
writer.  This means we use the Velox default which is to enable dictionary encoding for both integer
and string types.  The default in Presto is to disable dictionary encoding of integer types.

This change adds support for those configs to HiveConfig and plumbs them through HiveDataSink
down to the Writer. I modified the ColumnWriters to include this in their useDictionaryEncoding()
functions.

Reviewed By: Yuhta

Differential Revision: D59541486

fbshipit-source-id: 499de9355308647603ca58d09963941e6410dd13