Fix deadlocking parallel join build issue (#10407)
Summary:
Parallel hash table build will stuck if an off-thread (not on driver thread) allocations trigger arbitration. This is because off-thread arbitration does not have thread-local driver context to suspend the current driver thread. When arbitration goes along, it waits for the task to be fully suspended. This will never happen because of the above mentioned. And the query will go into a stuck state without resolving.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10407

Reviewed By: xiaoxmeng

Differential Revision: D59416016

Pulled By: tanjialiang

fbshipit-source-id: db85db2e518fe25a0e6b2b87c02885ba94fb77ee