Fix Resource Group State Info
As part of the memoization introduction, previous PR https://github.com/prestodb/presto/pull/22764 introduced
an issue where the endpoint returns the same value all the time when caching is disabled. Fixing it with this change
and introducing unit tests to cover both caching enabled and not enabled scenario to verify the result.