Fix improper template instantiation in ColumnWriterTest (#10441)
Summary:
The template instantiation when using complex types results in attempting to create a vector that is only implemented for use with simple types.

This can be seen when attempting to compile with Clang 17. For example the following is attempted to be instantiated: facebook::velox::SimpleVector<facebook::velox::Array<int>>::comparePrimitiveAsc

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10441

Reviewed By: pedroerp

Differential Revision: D59776879

Pulled By: mbasmanova

fbshipit-source-id: 7f9e12f80dc7b7d34b569d3e0d8ba9ecd0ddd7be