Add support for cache periodic full persistence
Currently, maxWriteRatio(AsyncDataCache::Options), the maximum ratio of the number of in-memory cache entries being written to the SSD cache over the total number of cache entries, is set to the default value of 0.7, meaning that we only persist 70% of cache entries to the SSD and upon server restart 30% of cache will be lost.

This change adds an optional periodic daemon to flush cache data to ssd. This can help to make sure all the in-memory data is flushed to ssd.