Enable Monolithic build (#9948)
Summary:
This PR adds an option that combines all targets into a single `velox` target and build a single unified `libvelox.a`. This is entirely optional and defaults to off. This will make it easier for downstream users to use Velox by just linking to `velox::velox` in their own cmake instead of having to look through our cmake which targets they need. (Once we have a config etc.).

I have yet to test the install functionality (it's strictly a bonus, the monolithic build is the goal for now).

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9948

Reviewed By: kgpai

Differential Revision: D59974408

Pulled By: kevinwilfong

fbshipit-source-id: 67459767d247a1264e1c5465c248586ac047ef4f