Optimize evaluation by reducing peeled vectors size (#10521)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10521

In case the peeled vector is sparsely selected (e.g. as a result of
filtering) and peeled, we still need to allocate and initialize/copy the memory
for unselected rows in result vector during evaluation, causing large
performance penalty in some queries.  Fix this by compacting the peeled vector
if the number of selected rows is smaller than 1/8 of the underlying base
vector.  A typical slow query due to this improved from 17.32 CPU days to 1.65
days, comparing to Java using 3.53 days.
bypass-github-export-checks

Reviewed By: oerling

Differential Revision: D60064934

fbshipit-source-id: 23ef4c590ebab55dc5551645fb705962e00c81a7