Timezone conversions based on seconds (#10563)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10563

Timezone conversions were done based on a timepoint<milliseconds>,
which is unecessary, adds possibility for overflow, and requires us to floor it
after the conversion. Simplifying the code and adding some more tests.

Now, only exporting a time_point<millisecond> may overflow, but not exporting
to time_point<second> which is what is needed for timezone conversions.

Reviewed By: kagamiori, mbasmanova

Differential Revision: D60189530

fbshipit-source-id: 0310fdabeb6441e9f341c140e2ee2e438a20080c