Fix assignments for equijoin on TimestampWithTimeZone
TimestampWithTimeZone can be represented in multiple equivalant ways.
This breaks and optimization introduced by f6c2252 because while
the values may be semantically equivalent, the contents may still
differ.  To support this, and future types that may end up needing this
support (such as JSON), a new flag is introduced to the type to indicate
whether or not the values of the type may support more than one
representation.