[native] Switch to Velox based signal handler for Linux
If Folly (used by Velox) is built with the correct symbolize dependencies
(need libunwind and libdwarf) then it can be used to handle signals and
produce a stack trace plus additional information such as the queryId.

On MacOS, however, libunwind is not available and it does not produce
a stack trace at all, not even an indication something bad had happened.
Therefore, on MacOS, the Google signal handler is used to produce a
stack trace.