Exclude map-type grouping keys for Spark aggregation fuzzer test (#10529)
Summary:
In this PR, option 'orderableGroupKeys' is added as a parameter of
AggregationFuzzer and WindowFuzzer. It is enabled for Spark aggregation fuzzer
test as Spark does not accept map-type grouping keys and distinct columns.
Method 'randOrderableType' taking 'scalarTypes' as parameter is added to
generate grouping keys for Spark.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10529

Reviewed By: pedroerp

Differential Revision: D60452458

Pulled By: Yuhta

fbshipit-source-id: e99dee935f8536ac182c62d13f2bf037fcb43b0e