Add isPathSupportedByRegisteredFileSystems api (#10603)
Summary:
We would like to add API isPathSupportedByRegisteredFileSystems to verify if
a given path is supported by registered file systems. This functionality is
necessary to ensure that in Gluten scan operators can fallback appropriately
when encountering paths not supported by the registered file systems in Velox.

Since this new API is intended for validation, it will not trigger any side
effects such as instantiating the file system or throwing exceptions.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10603

Reviewed By: pedroerp

Differential Revision: D60452574

Pulled By: Yuhta

fbshipit-source-id: 6ae81bb4b7da1e3aa42ec92135a9e6ad98b9a2f4