find_first and find_first_index have default null behavior (#10598)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10598

I verified in Presto Java find_first and find_first_index have default null behavior.  In Velox they do not
which leads to inconsistent behavior when actually evaluating the function could produce errors that
are skipped thanks to the optimizations that can be done when default null behavior is set.

The original code that set the non-default behavior included the following comment
    // find_first function is null preserving for the array argument, but
    // predicate expression may use other fields and may not preserve nulls in
    // these.
    // For example: find_first(array[1, 2, 3], x -> x > coalesce(a, 0)) should
    // not return null when 'a' is null.

If I'm reading this right, it's saying that null elements in the array may not produce null results in the
lambda.  AFAIK default null behavior isn't recursive so it only matters if the entire array is null.

There are already tests that verify if the array is NULL the result is NULL, I added tests that verify if
the optional start index argument is NULL the result is NULL.  I verified these passed before and after
my change, so this does not change the behavior of the function except in the presence of errors.

This is to address https://github.com/facebookincubator/velox/issues/10564

Reviewed By: pedroerp

Differential Revision: D60412963

fbshipit-source-id: 4f890fddc88e82363b27028f8e0d4a46f4c48107