[Iceberg] Use remainingPredicate in filter stats calc
This change combines the TableLayoutHandle's remaining predicate
and the pushed-down predicate when calculating statistics for
the call to the Iceberg connector's `getTableStatistics` in order to
attempt to get more accurate statistic estimates. The logic mimics
the Hive connector.

Explanation:

There were two issues with the previous implementation

1. Non-hive implementations never calculated statistics using the
FilterStatsCalculatorService.
2. The hive implementation did not consider the
`remainingPredicate` field when using the FilterStatsCalculatorService

This commit remedies both issues with some small refactoring and
changes to the arguments the connector passes to the
`FilterStatsCalculatorService`.

The FilterStatsCalculatorService attempts to refine the stats
calculations returned from a connector by considering the
full table statistics and any predicates applied to the scan in
order to reduce the total row count and other statistics like NDVs.

When filter pushdown is enabled on the iceberg connector, only
some parts of the filter can be pushed into the table scan. The rest
is represented by the layout handle's `remainingPredicate` field.

The Hive connector implementation creates a new binary
expression using an AND of the `remainingPredicate` and
`domainPredicate` fields. This change updates the implementation
to do the same.