Fix error message for invalid real cast
The error message for invalid cast from REAL ->
BIGINT/INTEGER/SMALLINT/TINYINT was previously returning the integer
bits underlying the float instead of the float value itself. This fixes
that to use the float value in the error message. For example, casting
a NaN value to bigint that was represented as 0xFFC00000 would fail
with the message "Unable to cast -4194304 to bigint".  Now the error
message will say "Unable to cast NaN to bigint".