Add option for admission control for filesystem resources (#10452)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10452

This change adds a generic admission controller class that can be
used for filesystem resources like read bytes in flight or number
of read requests in flight. It also provides a way to report stats
for resource usage, queued count, queued wait times by allowing the
client to specify a metric name.

Reviewed By: kevinwilfong, Yuhta

Differential Revision: D59643306

fbshipit-source-id: ddc4ea9273d3148c248cddf8a5543c5bf3df75bb