[native] fix openssl not found on Ubuntu
The OPENSSL_ROOT_DIR points to a path that does not exist.
This was removed on Velox. It doesn’t cause a problem on MacOS or CentOS but
on Ubuntu it will prevent the build by throwing

Could NOT find OpenSSL, try to set the path to OpenSSL root folder in the
  system variable OPENSSL_ROOT_DIR (missing: OPENSSL_CRYPTO_LIBRARY
  OPENSSL_INCLUDE_DIR)

Removing setting the OPENSSL_ROOT_DIR fixes the issue.