Set timestamp time zone in parquet writer (#10474)
Summary:
When we use Gluten's Parquet write to write timestamp data, we noticed that the timestamp data read out differs by a 7-hour time zone difference compared to vanilla Spark. We observed that Spark automatically adjusts the time zone to UTC when writing timestamp data. However, Velox's Parquet write function does not make such an adjustment, due to the fact that we did not set the time zone when creating the Arrow TimestampType. To address this issue, this PR will set the time zone in the Arrow TimestampType based on the kSessionTimezone configuration.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10474

Reviewed By: pedroerp

Differential Revision: D60623587

Pulled By: kagamiori

fbshipit-source-id: e10a53efe978e545ed37c5a4fe86328f2affe642