Optimize cross joins for single record build side (#10690)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10690

Turns out that having a single record on the build side is common as
users often cross-join relations with literals. Cross joins in Velox are
implemented using Nested Loop Join (NLJ). Adding an optimization for this
case where the single build record gets wrapped with entire probe batches, so
that probe can be processed one batch at a time.

Reviewed By: Yuhta

Differential Revision: D60967489

fbshipit-source-id: 49cf39e73b7d2e87201dbb34b8461e8e7c6ae069