Optimize cross joins with single build batch (#10695)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10695

As a follow up from:
.
https://github.com/facebookincubator/velox/pull/10690
.
We can also optimize the case of a cross join with a single build batch to
simply wrap probe and build inputs into a dictionary, then avoiding copying
and not producing outputs vectors that are too small.
.
Check in-code header comments for more details.

Reviewed By: Yuhta

Differential Revision: D61042584

fbshipit-source-id: b72fd0c139940ebf1eed90c1074ee7e308ba7396