Refactor try-catch clause for createOrcPageSource
Refactor the OrcBatchPageSourceFactory.createOrcPageSource method as it catches
raw java.lang.Exception which is not a recommended practice. A better practice
involves catching the specific Exception types that can be thrown in the try-catch
block. Kept theOrcDataSource object outside of a try-with-resources block as it is the
responsibility of the ConnectorPageSource to close both the reader and datasource.

Added an unit test to verify that any RuntimeException thrown in try block is still
mapped to a PrestoException and thrown as expected.