Re-Optimize cross joins with single build batch (#10726)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10726

As a follow up from:
.
https://github.com/facebookincubator/velox/pull/10690
.
We can also optimize the case of a cross join with a single build batch to
simply wrap probe and build inputs into a dictionary, then avoiding copying
and not producing outputs vectors that are too small.
.
Check in-code header comments for more details.
.
This is a re-submission of the PR
.
https://github.com/facebookincubator/velox/pull/10695
.
which was recently reverted. There is additional logic required to add this
optimization for joins other than cross-joins. Re-adding only for cross-joins
for now to unblock.

Will add it for non-cross-joins along with more tests as a follow up.

Reviewed By: Yuhta

Differential Revision: D61164691

fbshipit-source-id: 3f7dfadfdaad464f22d0dea45f556b17435fb7c8