Set hash_probe_finish_early_on_empty_build to false by default. (#10735)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10735

Keeping `hash_probe_finish_early_on_empty_build` to true causes queries to fail with `split source is closed` error. This is very hard to debug and in all deployment setup we set it to `false`. Making the default behavior to `false` to avoid confusing query errors for users.

Reviewed By: xiaoxmeng

Differential Revision: D61193165

fbshipit-source-id: 0555a87d2c298f092fd6040899978ad61c7bc624