Back out "Optimize evaluation by reducing peeled vectors size" (#10742)
Summary:
This reverses an optimization in peeling where a flat vector is
created as a peeled vector if the number of rows to peel is less
than 1/8 the size of the dictionary's alphabet, to prevent the
circulation of large peeled vectors that can cause large
intermediate allocations and affect memory and performance.

This resulted in a bug where the Common sub expression optimization,
that relies on the address of the input vector to re-use the cached
result, would get a different flat vector in its subsequent
invocation but would have the same memory address and assume it can
re-use the cached results. This would then cause wrong results.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10742

Original commit changeset: 23ef4c590eba

Original Phabricator Diff: D60064934

Reviewed By: kevinwilfong

Differential Revision: D61240010

fbshipit-source-id: d19eb30c33cdfbbd16a58a2ae315da3af1f85d20