Improve correctness of stddev and variance with partial aggregation
When merging varianceStates for partial aggregation, if the current
state has zero rows, use the values from the other state without doing
computation. This prevents introducing error due to imprecision in
floating point numbers.

Additionally, change the way we combine means. This ensures that we
do not introduce error due to imprecision in multiplication/division
when the delta is 0. I think it should in general improve the error
introduced by the mean computation, but I don't have a rigorous proof
or even experimental data for this.