Ensure flattening an array or map does not result in an oversized values vector (#10760)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10760

We recently found a bug in the transform lambda expression where the
capture of a lambda is an array and each row has more than 1000
elements. When inside the lambda eval, the capture is replicated by
wrapping it in a dictionary that aligns each row with each element
of the array that the transform is applied to. It then attempted to
flatten this wrapped vector, resulting in a segmentation fault.

This is a bandaid fix for this issue to ensure we fail instead of
crashing. A proper fix for transform lambda will be added later.

Reviewed By: Yuhta

Differential Revision: D61305363

fbshipit-source-id: aaa3180a2414a5dd36fc956ae771c75ebb1b1b39