Use ASSERT_STREQ when comparing C-strings (#10759)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10759

We currently use ASSERT_EQ to compare two strings as `const char*`
in type tests. ASSERT_EQ compares their pointers instead of the
string content. This has been working because the two operands are
string literals and the compilers usually keep only one copy of a
distinct string literal in a read-only memory space, so the two
operands point to the same address. This diff makes the check safer
by using ASSERT_STREQ that compares the string content.

Reviewed By: xiaoxmeng

Differential Revision: D61299100

fbshipit-source-id: 62899f3674475794d69317432d407a78940e002f