Fix filter on missing field (#10777)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10777

When there is a subfield filter on a missing field, we used to ignore
the filter, because we thought all constant fields are supplied by planner
(via `infoColumns`) and should not be filtered in execution level.  This is not true for
the null constant caused by missing field in file.  So we add the checks for
this case.

Reviewed By: tanjialiang

Differential Revision: D61473660

fbshipit-source-id: d72ecae6d13567e546c83fc4135b0d7414c20170