Release excessively reserved memory in HashBuild even if non-reclaimable (#10782)
Summary:
When hash build is under table building stage, we reserve excessive amount of memory to account for the worst case scenario duplicate rows for NextRowVector (i.e. we assume every row in build table has duplicates, which in most cases is not true). Other than let the query fail because the current stage is unreclaimable, we can perform a desperate try to release the unused reserved memory, giving the query a chance to succeed.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10782

Reviewed By: xiaoxmeng

Differential Revision: D61510068

Pulled By: tanjialiang

fbshipit-source-id: 1d62804d22ab11d08080e7cf872da0656cbd1010