Fix error on multiple inclusions of FindArrow CMake (#10723)
Summary:
The FindArrow file could be called more than one time, for example, if additional connectors require Arrow dependencies.

In that case the add_library call invoked a second time fails with the target already being defined and needs to be avoided. If the library was already registered so were the others and we don’t need to check explicitly.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10723

Reviewed By: xiaoxmeng

Differential Revision: D61394655

Pulled By: kgpai

fbshipit-source-id: 28d324b79fb6dd807076ef2ea9b195cd2da5d750