Add Spark query runner (#10357)
Summary:
The Spark Query Runner is a tool designed to facilitate the testing of Velox.
It achieves this by executing SQL queries on Apache Spark and returning results
as Velox data format, which allows the comparison of results between Velox and
Spark. It is designed to run against Spark-3.5.1.
https://github.com/facebookincubator/velox/pull/9559

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10357

Reviewed By: xiaoxmeng

Differential Revision: D60970781

Pulled By: kgpai

fbshipit-source-id: 8793a66975fccb0ed1fd7f73f72ceff5a5a7af05