Add shrink target adjustment mechanism in memory arbitration (#10769)
Summary:
Create a protection when shrinking capacity of a memory pool. The protection intends to give the memory pool more headroom after shinking so that the chance of the same memory pool needing to increase capacity in the future is slimmer. This is to reduce the future number of arbitrations.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10769

Reviewed By: xiaoxmeng

Differential Revision: D61381079

Pulled By: tanjialiang

fbshipit-source-id: b8bfc9e5b7fbab4d723db0c63f240db92d94ef8b