Fix null compacting for ARRAY and MAP selective column reader (#10805)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10805

For array/map column readers, we were using `resultNulls_` as nulls.
This is working correctly in most of the cases, except when during `read` call,
we decide to reuse reader nulls, but later there is another filter shrink the
row set, and cause that we need to compact the reader nulls into result nulls.
Although we set the bits in `resultNulls_` correctly, the `returnReaderNulls_` is
not reset in this case and we still use the reader nulls in result vector.

Fix this by reusing the same implementation in struct column reader on all
complex type column readers.  Also clean up the unused `compactComplexValues`
method.

Reviewed By: HuamengJiang

Differential Revision: D61620252

fbshipit-source-id: b5d35c7f94a622a8f9365e2fff7e89e441d113b0