Ensure inputs are alive before re-using the results of a common sub-expression (#10837)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10837

We recently encountered a bug (https://github.com/facebookincubator/velox/pull/10742) that caused incorrect results
when the inputs to a shared expression were replaced with a temporary
vector whose lifetime was limited to a single invocation of the
expression. This resulted in a new temporary input being passed to the
common sub-expression, where this vector had the same memory address
but differently mapped rows, and ended up re-using results from a
previous invocation. To address this issue, we have made a change to
ensure that before re-using the results, we verify that the inputs
have not expired.

Reviewed By: bikramSingh91

Differential Revision: D61739703

fbshipit-source-id: 52d4bfc9dd34a62ff3472dcfeb9d2c9f5d84ff4c