Terminate all Velox macro calls with semicolon
An upcoming change in Velox is going to enforce the use of a semicolon
in such statement-like macros.

Velox's current implementation of these macros, for which a trailing
semicolon at the call site is optional, is problematic because the
semicolons can trip the C++ compiler's `-Wextra-semi` lint in some
circumstances. The lint is technically unimpeachable because the macro
calls would have continued to compile with the caller's semicolon
removed.

The overwhelming majority of Velox macro calls (in Velox as well as
Presto) were already written with a trailing semicolon before this
commit.