Fix Server Active State Endpoint
During start of the cluster, ResourceGroupManagerConfigurationManager is initialized with InitializingConfigurationManager class.
And later configuration manager is configured when loadConfigurationManager method is called. But between this time period
GET `/v1/info/state` endpoint returns the server state as ACTIVE. And if the server receives any queries during the period, they all fails
with SERVER_STARTING_UP error.
In this PR, we are addressing that by providing a flag in ResourceGroupConfigurationManager that is been checked and if not true, `/v1/info/state`
endpoint return server status as INACTIVE.