Fix textfile ambiguous timestamps and different storage timezones
For TEXTFILE tables, when structural types (array, map, row) contain
timestamp columns, we weren't converting to the hive storage time zone
as we were doing for primitive types. Instead the timestamps would be
interpreted in the JVM timezone

The conversion to hive storage time zone had a secondary benefit (even
when the timezones were the same) of fixing the handling of ambiguous
timestamps. Ambiguous timestamps are local times that can have more than
one unixtime representation. It happens commonly during the fall DST
conversion where the hour from 1-2am repeats. Generally in Presto we use
the earlier of thetwo possible times when the unixtime is ambiguous. However,
the hive library we use for parsing textfiles uses the later time. The code
for adjusting the time based on the hive storage time zone has a secondary
benefit of correcting ambiguous timestamps to the earlier unixtime
representation.

This change fixes those two issues for the structural type code path.