Add SimplePlanFragment to the SPI
Add SimplePlanFragment to the SPI, as well as required dependencies
(Partitioning, PartitioningScheme, StageExecutionDescriptor,
PlanFragmeentId, PartititioningHandle).

The SimplePlanFragment contains all the critical fields from
PlanFragment, and will be used for plugin plancheckers on the entire
PlanFragment (validate native plan).

Some of the  methods from Partitioning and PartitioningScheme were moved
to a utility class PartitioningUtils.