Fix interpreting like expressions with escape
https://github.com/prestodb/presto/pull/23456 incorrectly changed the
invokation of the like function to use the unescaped pattern for a call
that expects the pattern to still contain the escape characters. As a
result, any escaped characters were treated as though they were not
escaped.  This affects certain queries when all arguments to the like
predicate are constants.