Extend expression fuzzer test to support decimal (#9149)
Summary:
ArgumentTypeFuzzer could be used to generate argument types when constructing a
decimal expression in the fuzzer test. However, given a result type, it is
unable to produce argument types that satisfy the necessary constraints. To
address this limitation, argument type generators for Presto and Spark decimal
functions have been added. In this PR, ExpressionFuzzer takes a map from
function name to an instance of the argument generator. Custom generators
provided by Presto and Spark are used in the expression fuzzer test to generate
argument types. Experimental fuzzer tests with decimal type enabled are added.

https://github.com/facebookincubator/velox/issues/1968

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9149

Reviewed By: kgpai

Differential Revision: D62193491

Pulled By: gggrace14

fbshipit-source-id: 6a0d935c4df4bfe4af605010b7df23a062fefc5a