[native] Replace fixed test worker port with ephemeral ports
Previously the listener ports for the native works in the E2E tests
was hard coded to be 1234 + worker count.
The change looks in the OS for an available ephemeral port
and uses this value when spawning the native workers.

The native worker must is deferring some configuration until the
port selection by the OS is known.