[native] Add native endpoint for Velox plan conversion
This adds an endpoint to the native Presto server that will convert
a Presto plan fragment to Velox. If the conversion is successful,
the server will send an ok response. If it fails, the server will
send an error response with a 422 status code as unprocessable.
The error message will contain a PlanConversionFailureInfo struct
with error type, code and message.

See also #23649
RFC: https://github.com/prestodb/rfcs/blob/main/RFC-0008-plan-checker.md