Extend assertEqualResults to compare floating-point columns in complex-type columns with epsilon (#10927)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10927

This diff addresses the limitation in https://github.com/facebookincubator/velox/issues/4481.

One limitation is still left that floating-point in map keys are currently not compared with
epsilon. This is because fuzzers can generate floating-point keys that are very close to
each other (with in epsilon), making one key incorrectly match another during comparison.

Reviewed By: xiaoxmeng

Differential Revision: D62222471

fbshipit-source-id: 4bba6ff01c61dc6f9c1a719cc3543b37b1ab84e8