Remove the shared string allocator use case (#11002)
Summary:
Remove the existing use cases on shared string allocator which cause complexity inside the row container.
The shared string allocator is not thread-safe so it is better not to share with the other components. The
memory overhead within a shared string allocator is pretty small which is just one or two pages. The
shared string allocator is previously used for in-memory spilling operation which has been removed.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11002

bypass-github-export-checks

Reviewed By: amitkdutta, oerling

Differential Revision: D62609465

fbshipit-source-id: 44fafb9f943a2b1567e745abec21844df880cd85