Add check for overlapping ranges of ARRAY and MAP vectors (#10960)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10960

We don't allow overlapping ranges in ARRAY and MAP vectors.  However this is not clear in the code, so we add a method for user to check that.

In the future we will also add the check to `BaseVector::validate()` and `ArrayVectorBase` constructor to make it clear it is not allowed.

Reviewed By: xiaoxmeng

Differential Revision: D62212238

fbshipit-source-id: 6ae0eff4f76ad638de3994fc6bff44d7ba784838