Fix hash build not able to reclaim in finishHashBuild (#11016)
Summary:
This is a regression caused by previous change. When hash build is in finishHashBuild() function, it ought to be reclaimable while calling ensureTableFits(). But ensureTableFits was called after spiller_ and table_ ownerships have been transferred, which caused the whole operator to be in a not able to reclaim state. This PR fixed the issue and added tests to avoid similar issues from happening again.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11016

Reviewed By: xiaoxmeng

Differential Revision: D62783315

Pulled By: tanjialiang

fbshipit-source-id: f9d4cc1c1e07fd64b3f006ee231d967a5d033493