Support UNSUPPORTED_INPUT_UNCATCHABLE VeloxRuntimeError in expression fuzzer (#10993)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/10993

We recently made cast-varchar-to-integral throw VeloxRuntimeError on unicode
inputs and made date-time functions throw VeloxRuntimeError when timestamps
are out of supported range. The expression fuzzer tests have been failing due to
these changes because VeloxRuntimeError is not allowed in fuzzer. This diff
introduces a new ErrorCode `UNSUPPORTED_INPUT_UNCATCHABLE` to VeloxRuntimeError that is
ignored in fuzzers, and makes the cast and date-time functions throw errors with
this error code.

Reviewed By: xiaoxmeng, kgpai

Differential Revision: D62612232

fbshipit-source-id: 334d615555a12c47d9b2219dca1eb382edcfabf3