Allow Custom Types to provide custom implementations of compare and hash functions. (#11021)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11021

**Problem**:
Today, when a Custom Type extends an existing physical type, it inherits the comparison and hashing
semantics of the native type. This can lead to undesired semantics. For example, we've seen for
TimestampWithTimeZone, it inherits its equality and hashing from int64_t, but in order to match
Presto semantics we need times at different time zones that are equivalent when converted to UTC
to be considered equal generally.

**Solution**:
This change attempts to address this:
1) I added a boolean parameter `providesCustomComparison` to Type which is set in the
constructor. Custom types that want to provide their own implementation of compare & hash set this
to true in their constructor.
2) I added virtual functions `compare` and `hash` to TypeBase. Custom types that want to provide
their own implementation of compare & hash implement them by overriding these functions. These
will be invoked instead of the native type's compare and hash functions if
`providesCustomComparison` is set. By default these throw, as they should not be invoked if
`providesCustomComparison` is not set, and if `providesCustomComparison` is set they must be
overridden.

I will update the places where we do comparisons/hash to make use of this in diffs later in the stack.

**Limitations**:
For now I've limited this to primitive, fixed width types.  Primitive most non-primitive types do not
have a well defined NativeType, so how to implement custom compare & hash functions is not
immediately clear (at least to me). Fixed width because in some places, e.g. ContainerRowSerde,
they take advantage of the fact that compare does not need the data to be contiguous in memory to
avoid copying the data, and without a motivating use case I didn't want to complicate this logic.

**Alternatives Considered**:
* Add a new TypeKind for new custom types that need custom comparisons. This requires changes
to Velox core code for custom types form customers, which removes a lot of the value of custom
types.
* Just have virtual `compare` and `hash` functions (no `providesCustomComparison`). The
conditional already adds some overhead, a dynamic call in the general case would be prohibitively
expensive.
* Add a `TypeWithCustomComparison` base class (or something like that) so that we don't need
to make the `compare` and `hash` functions only available in custom types that don't need it. This
would mean casting to `TypeWithCustomComparison<TypeKind>` in places where we need to call
the custom functions. This would be less safe than casting to `TypeBase<TypeKind>` (since all types
inherit from some `TypeBase<TypeKind>`) and require a dynamic_cast which would be more
expensive.
* Somehow work this into `CustomTypeFactories`. This would require a map lookup which would be
even more expensive than a dynamic call.

Reviewed By: Yuhta

Differential Revision: D62900789

fbshipit-source-id: a7c67c0cf716eebf829bdc3e9761126f337fc8db