Fix internal build of velox_expression_fuzzer_test (#11044)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11044

A previous change happened to break the internal build of expression_fuzzer_test, but the
broken build was not caught at diff-time internally because the expression_fuzzer_test
target was disabled. This target is useful for debugging. This diff fixes the broken build.
Plus, since we only want expression_fuzzer_test to be built but not run at diff time, this diff
changes the target type to cpp_binary and removes the GTest logic that was required to run
it as cpp_unittest before.

Reviewed By: kevinwilfong, kgpai

Differential Revision: D63042420

fbshipit-source-id: 5ac95c5d0692803a0e95039c0954b97f4fccda58