Allow pushdown of dynamic filters through HashAggregation (#10988)
Summary:
Currently, in HashAggregation, only when the aggregation is distinct,
the channel of the group key is added to the identity projections. This
results in a situation where, when the group key matches the join key,
the dynamic filters generated by HashProbe cannot be pushed down through
HashAggregation. This pr supports dynamic filters push down through
HashAggregation when group key and join key is the same.

This optimization applies to plans like:

```
- HashProbe(k1 = k2,...)
  - HashAggregation(k1, sum(x))
    - TableScan
```

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10988

Reviewed By: kevinwilfong

Differential Revision: D63035508

Pulled By: Yuhta

fbshipit-source-id: 7fb336adb17154c54658a7db518a94b79157efce