CMake: Only set the variable if it has not been defined (#11040)
Summary:
This enables variables to be set via `-Dvar_name=foo` on the command-line. Traditionally that would override an environment variable. This is more of a minor adjustment to the behavior and it will not break existing scripts using environment variables but a CMake cache variable would take precedence if both were supplied.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11040

Reviewed By: xiaoxmeng

Differential Revision: D63138389

Pulled By: Yuhta

fbshipit-source-id: 3deac39616600f808195df8e94ef3b711fb7f9bb