Support custom comparison in Vectors (#11022)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11022

Building on https://github.com/facebookincubator/velox/pull/11021 this adds support for custom
comparison functions provided by custom types in the compare and hash functions in the
BaseVector implementations.

Since this is only supported for primitive types the changes are largely contained in just the
SimpleVector class.

Note that this will introduce a regression in the general performance of compare and hash since we
need to check another condition on type->providesCustomComparison(). Unfortunately, I think
some sort of runtime check there is unavoidable.

Reviewed By: Yuhta

Differential Revision: D62901338

fbshipit-source-id: 8cca5a5b5f739bac8e9c276100a38ffde2c9aeb6