CMake: Make the env variable precedence by default in set_with_default (#11070)
Summary:
After the modification in https://github.com/facebookincubator/velox/issues/11040, if the `var_name` is already defined, the `set_with_default` method will not override it with the env variable. However, before that, the value from the env will be prioritized for defining `var_name`. The design goal of the `resolve_dependency_url` method is to allow the overriding of values defined in `VELOX_xxx_SOURCE_URL` with env variables, which enables specifying `VELOX_xxx_SOURCE_URL` to a local file to reduce dependency download time and achieve offline dependency compilation.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11070

Reviewed By: DanielHunte

Differential Revision: D63344157

Pulled By: Yuhta

fbshipit-source-id: 3740d3a77f12f096f7c7c9d5460b005b3a9815e0