Make concat function throw when there are less than 2 arguments (#11076)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11076

Currently, the velox concat function handles having a single argument by returning the argument as the result. Presto throws an exception, so velox should as well.

This change was made before and had to be backed out due to failing XStream pipelines. XStream A/B test in ga and f3 have been verified to succeed this time.

Reviewed By: kevinwilfong

Differential Revision: D63271129

fbshipit-source-id: 60afd7821aa8b49707a9d6c8f0b5421a006a60b6