Fix null value check for kRange frame column (#11075)
Summary:
The columns in `WindowPartition::data_` are reordered to start with partition-by keys and order-by keys, whereas the columns in `WindowPartition::columns_` are in the same order as the Window operator. The variable `orderByColumn` contains the index of order-by key in the reordered columns, so this should be accessed from `data_` and not `columns_`. A unit test is added to ensure an exception is thrown when NULL values in order-by column and frame column do not match.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11075

Reviewed By: amitkdutta

Differential Revision: D63404094

Pulled By: kagamiori

fbshipit-source-id: cae3713c3c6c945001af8c2930ba08a34d574f67