Support custom comparison in Presto's IN function (#11032)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11032

Building on https://github.com/facebookincubator/velox/pull/11021 this adds support for custom
comparison functions provided by custom types in Presto's IN function.

I was able to reuse the ComplexTypeInPredicate and the support for custom comparisons already
present in BaseVector.

This diff is largely just renaming ComplexTypeInPredicate to VectorSetInPredicate (to clarify it's not
just for complex types anymore) and if statement to identify the case where
providesCustomComparison() is true for the element type (and of course updating the tests).

Making TimestampWithTimeZone a special case of bigint (comparing the millis) in the future might
give a performance boost if this shows up as a bottleneck.

Reviewed By: xiaoxmeng

Differential Revision: D62994557

fbshipit-source-id: 82d3eb2c3d24118f7f555b3f739810c91c58fc32