Allow task writer counts to be non-power of two for native engine
task_partitioned_writer_count is used to configure the number of drivers at worker size
which doesn't have to be power of two. In Prestissimo we need to configure this to avoid
the number of table writer workers is a multiple of the number of table writer threads per
worker. As Presto java engine still requires table writer threads to be power of two so
keep the check for non-native engine scenario.